backend/requirements.txt
fastapi
uvicorn
pyyaml
requests
pydantic
